refactor of DumpIndex -- move to index package
