+++ src/main/java/io/anserini/index/README.md
@@ -1,32 +1,32 @@
 DumpIndex provides an easy way to print various information of the index via CLI
 
-```-index```
+`-index`
 
 The path of the index
 
-```-s```
+`-s`
 
 Print statistics for the Repository: total docs count, total terms count, etc.
 
-```-t```
+`-t`
 
 Print term info: stemmed, total counts, doc counts
 
-```-di```
+`-di`
 
 Print the internal document IDs of documents
 
-```-dn```
+`-dn`
 
 Print the text representation of a document ID
 The document ID is the internal document ID starting from 1
 
-```-dt```
+`-dt`
 
 Print the text of a document (only if the raw documents are stored)
 The document ID is the internal document ID starting from 1
 
-```-dv```
+`-dv`
 
 Print the document vector of a document (only if the document vectors are stored)
 The document ID is the internal document ID starting from 1
